requirements.txt
Django~=3.2.4
django-filter~=2.4.0
django-debug-toolbar~=3.2.1
djangorestframework>=3.12.4
djangorestframework-simplejwt
drf-nested-routers~=0.93.3
phonenumbers==8.12.25
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
